refactor: db/routes structural cleanup
