+++ server/src/middleware/rateLimiter.js
@@ -24,11 +24,6 @@ export function createRateLimiter({ windowMs, maxRequests, keyFn }) {
   cleanupHandle.unref?.();
 
   const middleware = async (c, next) => {
-    if (process.env.DISABLE_RATE_LIMIT === "true") {
-      await next();
-      return;
-    }
-
     const key = resolveKey(c);
     const now = Date.now();
     const timestamps = store.get(key) || [];
@@ -42,6 +37,5 @@ export function createRateLimiter({ windowMs, maxRequests, keyFn }) {
     store.set(key, recent);
     await next();
   };
-  middleware.stop = () => clearInterval(cleanupHandle);
   return middleware;
 }
