+++ server/src/routes/auth.js
@@ -26,9 +26,6 @@ export function _resetRateLimits() {
 }
 
 function checkRateLimit(bucketKey) {
-  if (process.env.DISABLE_RATE_LIMIT === "true") {
-    return true;
-  }
   return dbUtils.checkRateLimit(bucketKey, RATE_LIMIT.WINDOW_MS, RATE_LIMIT.MAX_ATTEMPTS);
 }
